requirements.txt
selenium
pandas
pywin32
pillow
requests